+++ app/schemas/purchase.py
@@ -4,7 +4,7 @@ from datetime import datetime
 
 class PurchaseRequest(BaseModel):
     clientId: int = Field(..., examples=[1])
-    amount: Decimal = Field(..., gt=0, examples=[250])
+    amount: float = Field(..., gt=0, examples=[250])
     currency: str = Field(..., min_length=3, max_length=3, examples=["USD"])
     purchaseDate: datetime = Field(..., examples=["2025-09-20T14:30:00Z"])
     purchaseCountry: str = Field(..., examples=["Francia"])
