+++ app/utils/monetary.py
@@ -1,5 +1,5 @@
 from decimal import Decimal, ROUND_HALF_UP
 
-def money(x: Decimal) -> Decimal:
+def money(x: float) -> float:
     # Redondeo estándar a 2 decimales
-    return (x.quantize(Decimal("0.01"), rounding=ROUND_HALF_UP))
+    return (x.quantize(float("0.01"), rounding=ROUND_HALF_UP))
